backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.5
pydantic==2.5.0
python-multipart==0.0.6
fastapi-cors==0.0.6
# PDF processing
pdfplumber==0.10.3
pymupdf==1.23.8
# Database
psycopg2-binary==2.9.9
# Text processing
tiktoken==0.5.2
numpy==1.24.3
# PDF generation (for example PDF)
reportlab==4.0.7

